requirements.txt
streamlit
crewai
crewai-tools
openai
python-dotenv
pandas
numpy
matplotlib
seaborn
scikit-learn
shap